requirements.txt
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn==21.2.0
whitenoise==6.6.0 